Change: Pipes.VMCP.Marionette.sendMarionetteMsg> More 'Properly' bundling [WIP]
Blendshapes and bone transformation should be bundled, and others are not.

I asked about how I have to bundle messages on EVMC4U discord server,
and learnt that Blend shapes and bone transformations are bundled in
many cases (although it's not explicitly defined)

---
This is still in development, as it seems like it doesn't send message
right now.